requirements.txt
flask==2.0.1
google-cloud-storage>=1.33.0
gunicorn
python-dotenv